fix merge conflicts re #325
