+++ mslice/app/mainwindow.py
@@ -3,8 +3,9 @@ from __future__ import (absolute_import, division, print_function)
 from mslice.util.qt.QtWidgets import QApplication, QMainWindow, QLabel, QMenu
 
 from mslice.presenters.main_presenter import MainPresenter
+from mslice.presenters.slice_plotter_presenter import SlicePlotterPresenter
 from mslice.util.qt import load_ui
-from mslice.views.mainview import MainView
+from mslice.views.interfaces.mainview import MainView
 from mslice.widgets.ipythonconsole.ipython_widget import IPythonWidget
 from mslice.widgets.workspacemanager import TAB_2D, TAB_EVENT, TAB_HISTO, TAB_NONPSD
 from mslice.widgets.workspacemanager.command import Command as ws_command
@@ -40,12 +41,14 @@ class MainWindow(MainView, QMainWindow):
             self.buttons_to_enable[TAB_HISTO] += [self.btnSaveToADS]
 
         self.workspace_presenter = self.wgtWorkspacemanager.get_presenter()
-        self.dataloader_presenter = self.data_loading.get_presenter()
-        self.slice_presenter = self.wgtSlice.get_presenter()
-        self.powder_presenter = self.wgtPowder.get_presenter()
+        dataloader_presenter = self.data_loading.get_presenter()
+        slice_plotter_presenter = SlicePlotterPresenter()
+        slice_widget_presenter = self.wgtSlice.get_presenter()
+        slice_widget_presenter.set_slice_plotter_presenter(slice_plotter_presenter)
+        powder_presenter = self.wgtPowder.get_presenter()
         self.cut_presenter = self.wgtCut.get_presenter()
-        self._presenter = MainPresenter(self, self.workspace_presenter, self.dataloader_presenter,
-                                        self.slice_presenter, self.powder_presenter, self.cut_presenter)
+        self._presenter = MainPresenter(self, self.workspace_presenter, dataloader_presenter,
+                                        slice_widget_presenter, powder_presenter, self.cut_presenter, slice_plotter_presenter)
 
         self.wgtWorkspacemanager.tab_changed.connect(self.ws_tab_changed)
         self.setup_save()
