+++ mslice/plotting/plot_window/plot_figure_manager.py
@@ -80,12 +80,12 @@ class PlotFigureManagerQT(QtCore.QObject):
     def flag_as_current(self):
         self.window.flag_as_current()
 
-    def add_slice_plot(self, slice_plotter, workspace):
+    def add_slice_plot(self, slice_plotter_presenter, workspace):
         if self._plot_handler is None:
             self.move_window(-self.window.width() / 2, 0)
         else:
             self._plot_handler.disconnect(self.window)
-        self._plot_handler = SlicePlot(self, slice_plotter, workspace)
+        self._plot_handler = SlicePlot(self, slice_plotter_presenter, workspace)
 
     def add_cut_plot(self, cut_plotter, workspace):
         if self._plot_handler is None:
