+++ mslice/tests/main_presenter_test.py
@@ -4,7 +4,7 @@ import unittest
 
 from mslice.presenters.main_presenter import MainPresenter
 from mslice.presenters.interfaces.workspace_manager_presenter import WorkspaceManagerPresenterInterface
-from mslice.views.mainview import MainView
+from mslice.views.interfaces.mainview import MainView
 
 SELECTED_WORKSPACES = ['a', 'b', 'c']
