+++ mslice/views/interfaces/slice_view.py
@@ -1,7 +1,7 @@
 from __future__ import (absolute_import, division, print_function)
 
 
-class SlicePlotterView:
+class SliceView:
     error_occurred = None
     busy = None
