+++ mslice/views/interfaces/workspace_view.py
