+++ mslice/widgets/projection/powder/powder.py
@@ -11,7 +11,7 @@ from mslice.util.qt.QtWidgets import QWidget, QMessageBox
 from mslice.models.projection.powder.mantid_projection_calculator import MantidProjectionCalculator
 from mslice.presenters.powder_projection_presenter import PowderProjectionPresenter
 from mslice.util.qt import load_ui
-from mslice.views.powder_projection_view import PowderView
+from mslice.views.interfaces.powder_projection_view import PowderView
 from .command import Command
